Add long int in line 12

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -9,7 +9,7 @@
 
 void print_number(int n)
 {
-	int power10 = 1, digit;
+	long int power10 = 1, digit;
 
 	if (n < 0)
 	{
